+++ Network/Packets/Implementation/PlayerPositionPacket.cs
@@ -10,7 +10,7 @@ using UnityEngine;
 namespace AMP.Network.Packets.Implementation {
     [PacketDefinition(false, (byte)PacketType.PLAYER_POSITION)]
     public class PlayerPositionPacket : AMPPacket {
-        [SyncedVar]       public long    playerId;
+        [SyncedVar]       public int     playerId;
 
         [SyncedVar(true)] public Vector3 handLeftPos;
         [SyncedVar(true)] public Vector3 handLeftRot;
@@ -26,7 +26,7 @@ namespace AMP.Network.Packets.Implementation {
 
         public PlayerPositionPacket() { }
 
-        public PlayerPositionPacket(long playerId, Vector3 handLeftPos, Vector3 handLeftRot, Vector3 handRightPos, Vector3 handRightRot, Vector3 headPos, Vector3 headRot, Vector3 playerPos, float playerRot) {
+        public PlayerPositionPacket(int playerId, Vector3 handLeftPos, Vector3 handLeftRot, Vector3 handRightPos, Vector3 handRightRot, Vector3 headPos, Vector3 headRot, Vector3 playerPos, float playerRot) {
             this.playerId     = playerId;
             this.handLeftPos  = handLeftPos;
             this.handLeftRot  = handLeftRot;
